feat: improve the code structure for easy maintenance
change FileBrowserClient code more clear

feat #1
