+++ file_browser_client/header.go
@@ -4,15 +4,15 @@ const (
 	baseUserAgent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/109.0.0.0 Safari/537.36"
 )
 
-var baseHeader map[string]string
+var baseHeaderSave map[string]string
 
 func BaseHeader() map[string]string {
-	if baseHeader != nil {
-		return baseHeader
+	if baseHeaderSave != nil {
+		return baseHeaderSave
 	}
 	header := map[string]string{
 		"User-Agent": baseUserAgent + " github.com/sinlov/filebrowser-client",
 	}
-	baseHeader = header
-	return baseHeader
+	baseHeaderSave = header
+	return baseHeaderSave
 }
