--- file_browser_client/resourcesDefine.go
@@ -1,13 +1,0 @@
-package file_browser_client
-
-type (
-	ResourcePostFile struct {
-		RemoteFilePath string
-		LocalFilePath  string
-	}
-
-	ResourcePostDirectory struct {
-		RemoteDirectoryPath string
-		LocalDirectoryPath  string
-	}
-)
