+++ go.mod
@@ -2,19 +2,18 @@ module github.com/sinlov/filebrowser-client
 
 go 1.17
 
-require github.com/stretchr/testify v1.8.0
-
 require (
-	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
-	github.com/russross/blackfriday/v2 v2.1.0 // indirect
-	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
-	gopkg.in/yaml.v2 v2.4.0 // indirect
+	github.com/monaco-io/request v1.0.16
+	github.com/stretchr/testify v1.9.0
+	github.com/urfave/cli/v2 v2.27.1
 )
 
 require (
+	github.com/cpuguy83/go-md2man/v2 v2.0.4 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/monaco-io/request v1.0.15
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/urfave/cli/v2 v2.23.7
+	github.com/russross/blackfriday/v2 v2.1.0 // indirect
+	github.com/xrash/smetrics v0.0.0-20240312152122-5f08fbb34913 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
